--- a/datafusion/core/src/physical_plan/aggregates/topk/hash_table.rs
+++ b/datafusion/core/src/physical_plan/aggregates/topk/hash_table.rs
@@ -335,7 +335,6 @@ impl<ID: KeyType> HashTableItem<ID> {
     }
 }
 
-#[allow(dead_code)]
 #[cfg(test)]
 fn map_print<ID: KeyType>(map: &RawTable<HashTableItem<ID>>) {
     use itertools::Itertools;