… datafusion/core/Cargo.toml
… datafusion/core/src/physical_plan/aggregates/topk/mod.rs
… datafusion/core/src/physical_plan/aggregates/topk_stream.rs
… datafusion/core/tests/sql/group_by.rs
… datafusion/optimizer/tests/optimizer_integration.rs
… datafusion/physical-expr/src/aggregate/utils.rs
… datafusion/sqllogictest/test_files/aggregate.slt
… datafusion/sqllogictest/test_files/explain.slt
… datafusion/sqllogictest/test_files/information_schema.slt
… docs/source/user-guide/configs.md