## Which issue does this PR close?

Closes #7191.

## Rationale for this change

Described in issue.

## What changes are included in this PR?

1. A new `GroupedTopKAggregateStream` aggregation
2. A new `limit` property on `AggregateExec`
3. An optimizer rule to copy the limit from the `SortExec` if applicable

## Are these changes tested?

- [x] `AggregateExec` now prints `lim=X` if there's a limit, and I added some tests to assert this
- [x] unit tests for the Map & Heap
- [x] need to A/B test with and without this rule enabled
- [x] need to add criterion benchmarks
- [x] need to performance optimize until it is at least no slower than the existing approach
- [x] `sqllogictest`s to compare to existing functionality

## Are there any user-facing changes?

1. Some Top K queries should not crash
2. ~~I probably broke other things so this is a draft~~ All the existing tests now pass

## Notes

Concerns to address:

- [x] ~~the `OwnedRow` code is not columnar, vectorized, etc~~ most queries will use a single column
- [x] ~~use the existing Acculumators?~~ not required since this is only min/max
- [x] ~~filters are not yet applied~~ unsupported edge case for now
- [x] NULLs are supported
- [ ] this should be a whole new `Exec` node, not just a new `Stream` type?
- [x] ~~key types other than `String`~~ now supports String + all primitive keys
- [x] replace `TreeMap` with custom index-based heap

## Out of scope

- [ ] handle multiple keys & values with `OwnedRow`
- [ ] performance boost by moving from index-heap to RawPointer heap?