@alamb 2023-08-05
The more I think about this code / approach the more I like it ❤️  -- I spent some time writing up how I think this basic strategy can be applied to all the various TopK type queries at https://github.com/apache/arrow-datafusion/issues/7198#issuecomment-1666547553

I think my writeup assumes a slightly different packaging / deciding how to invoke this operator, but the basic idea I think is the same. 

Thank you for sparking this @avantgardnerio