@alamb 2023-08-07
> Would anyone be able to provide advice on debugging sql logic tests? This error doesn't seem very informative.. I'd expect to see more of a diff than this:?

The docs are here: 
https://github.com/apache/arrow-datafusion/tree/main/datafusion/core/tests/sqllogictests

using 

```shell
cargo test --test sqllogictests -- --complete
```

Would likely save you time

I believe that diff says a new line was added to the explain plan (which makes sense if you have added a new optimizer pass)