@Dandandan 2023-08-10
@avantgardnerio seems best to profile it ATM and see where the most time is spent