@avantgardnerio 2023-08-10
Relevant trace: 
<img width="1051" alt="Screenshot 2023-08-10 at 8 58 00 AM" src="https://github.com/apache/arrow-datafusion/assets/3855243/440e9cf1-1d69-44d8-b6f4-e33a9942e7d1">

Not one big expense, just lots of the little ones we all expected.