@avantgardnerio 2023-08-10
Initial results (baseline) for `should_perform_well_on_realistic_data` on `bbc3b2f62efec3d41d0e48bf05988e9836d643a6`:

1. debug: `Aggregated 100000 records in 363.2565ms`
2. release: `Aggregated 100000 records in 41.484667ms`