@avantgardnerio 2023-08-10
> some high, fixed cost to running this stream

I think that cost is converting the columns to rows, which appears to be about 30% of the overall cost of this new stream.