@alamb 2023-08-10
FWIW I am working on a generic TopK operator (for `ORDER BY / LIMIT`) in  https://github.com/apache/arrow-datafusion/pull/7250

One of the things I will eventually need there a Max/Max Heap that has the properties you are describing, so I look forward to seeing how you do it too (depending on the timeline I may try and hack up a version for myself as well)

BTW I think to truly achieve no allocations we will need to extend the arrow_row API to reuse existing `Row` (owned rows). But I have a lot of work to do before I am sure the owned row is the problem